feat: Add cost mode and period flag parsing

+++ app/cmd/destroy.go
@@ -15,7 +15,11 @@ var destroyCmd = &cobra.Command{
 	Long:  `Run terraform destroy and show changes in Pluralith`,
 	Run: func(cmd *cobra.Command, args []string) {
 		tfArgs := terraform.ConstructTerraformArgs(cmd.Flags())
-		costArgs := cost.ConstructInfracostArgs(cmd.Flags())
+		costArgs, costErr := cost.ConstructInfracostArgs(cmd.Flags())
+		if costErr != nil {
+			fmt.Println(costErr)
+			return
+		}
 
 		if destroyErr := terraform.RunTerraform("destroy", tfArgs, costArgs); destroyErr != nil {
 			fmt.Println(destroyErr)
